feat: hook 분리
